Implemented resizable base disks

+++ .build/psake.ps1
@@ -76,26 +76,30 @@ Task Test -Depends Init {
     "`n`tSTATUS: Testing with PowerShell $PSVersion"
 
     # Ensure recent Pester version is actually used
-    Import-Module -Name Pester -MinimumVersion 4.0.0 -Force
-
-    # Gather test results. Store them in a variable and file
-    $TestResults = Invoke-Pester -Path $ProjectRoot\Tests -PassThru -OutputFormat NUnitXml -OutputFile "$ProjectRoot\$TestFile"
-
-    # In Appveyor?  Upload our tests! #Abstract this into a function?
-    If ($ENV:BHBuildSystem -eq 'AppVeyor')
+    if (-not $IsLinux)
     {
-        (New-Object 'System.Net.WebClient').UploadFile(
-            "https://ci.appveyor.com/api/testresults/nunit/$($env:APPVEYOR_JOB_ID)",
-            "$ProjectRoot\$TestFile" )
-    }
+        Import-Module -Name Pester -MinimumVersion 5.0.0 -Force
 
-    Remove-Item "$ProjectRoot\$TestFile" -Force -ErrorAction SilentlyContinue
+        # Gather test results. Store them in a variable and file
+        $TestResults = Invoke-Pester -Path $ProjectRoot\Tests | ConvertTo-NUnitReport -ErrorAction SilentlyContinue
+        $TestResults.Save("$ProjectRoot\$TestFile")
 
-    # Failed tests?
-    # Need to tell psake or it will proceed to the deployment. Danger!
-    if ($TestResults.FailedCount -gt 0)
-    {
-        throw "Failed '$($TestResults.FailedCount)' tests, build failed"
+        # In Appveyor?  Upload our tests! #Abstract this into a function?
+        If ($ENV:BHBuildSystem -eq 'AppVeyor')
+        {
+            (New-Object 'System.Net.WebClient').UploadFile(
+                "https://ci.appveyor.com/api/testresults/nunit/$($env:APPVEYOR_JOB_ID)",
+                "$ProjectRoot\$TestFile" )
+        }
+
+        Remove-Item "$ProjectRoot\$TestFile" -Force -ErrorAction SilentlyContinue
+
+        # Failed tests?
+        # Need to tell psake or it will proceed to the deployment. Danger!
+        if ($TestResults.FailedCount -gt 0)
+        {
+            throw "Failed '$($TestResults.FailedCount)' tests, build failed"
+        }
     }
     "`n"
 }
